+++ src/kernel/mediatek/mt8163/4.9/drivers/misc/mediatek/cmdq/v2/inc/cmdq_sec_dr/cmdq_sec_iwc_common.h
@@ -73,7 +73,7 @@ struct iwcCmdqAddrMetadata_t {
 	/* [IN] addr handle type*/
 	uint32_t type;
 	/* [IN]_h, secure address handle */
-	uint32_t baseHandle;
+	uint64_t baseHandle;
 	/* [IN]_b, block offset from handle(PA) to current block(plane) */
 	uint32_t blockOffset;
 	/* [IN]_b, buffser offset to secure handle */
